requirements.txt
requests
beautifulsoup4
wordcloud
matplotlib
python-docx
openai